spostate funzioni di utilità da veicolo.c a utile_veicolo.c
Separata la logica di supporto relativa ai veicoli in un nuovo file `utile_veicolo.c`
Migliorato l'information hiding e la modularità del codice
Aggiornati i riferimenti nei file header se necessario
