+++ veicolo.c
@@ -213,119 +213,29 @@ ptr_prenotazione prendi_prenotazioni(ptr_veicolo ve)
 
 
 /*
- Funzione: stampa_veicolo
- ------------------------
- Stampa a video le informazioni del veicolo in formato leggibile.
+ Funzione: imposta_stato
+ -----------------------
+ Imposta lo stato di un veicolo.
 
  Parametri:
-   ve: puntatore al veicolo da stampare
-
-  Pre-condizione:
-   Il puntatore ve deve essere valido (non NULL)
-
- Post-condizione:
-   Nessuna modifica alla struttura veicolo
-
- Effetti:
-   Stampa le principali informazioni del veicolo su standard output.
-*/
-void stampa_veicolo(ptr_veicolo ve)
-{
-    if(ve){
-        printf(" %s  %s [%s] \n\tposizione: %s\t\t(%s)", ve->marca, ve->modello, ve->targa, ve->posizione, ve->stato);
-    }
-}
-
-
-/*
-Funzione: carica_veicoli_da_file
- -------------------------------
- Carica i veicoli da un file e li inserisce in una lista.
-
- Parametri:
-   nome_file: stringa con il nome del file da leggere
-   l: puntatore alla lista di veicoli dove inserire i dati
-
-  Pre-condizione:
-   Il file deve essere formattato correttamente con righe contenenti marca, modello, targa e posizione
-
- Post-condizione:
-   Tutti i veicoli letti dal file vengono inseriti nella lista l
-
- Effetti:
-   Legge i dati da file e li aggiunge alla lista.
-   Se il file non è accessibile, il programma termina con errore.
-*/
-void carica_veicoli_da_file(const char *nome_file, ptr_lista l)
-{
-    FILE *file = fopen(nome_file, "r");
-    if(!file){
-        printf("!!!VEICOLI NON DISPONIBILI!!!");
-        exit(1);
-    }
-
-    char marca[30], modello[40], targa[8], posizione[60];
-    while(fscanf(file, "%30s %40s %8s %60s", marca, modello, targa, posizione) == 4){
-        inserisci_veicolo_lista(l, inizia_veicolo(marca, modello, targa, posizione));
-    }
-}
-
-
-/*
- Funzione: aggiorna_stato_veicolo
- --------------------------------
- Aggiorna lo stato di disponibilità di un veicolo in base alle prenotazioni.
-
- Parametri:
-   ve: puntatore al veicolo di cui aggiornare lo stato
-
-  Pre-condizione:
-   ve deve essere un puntatore valido a una struttura veicolo
-
- Post-condizione:
-   Il campo 'stato' del veicolo viene aggiornato a "disponibile" se il veicolo è disponibile oggi,
-   altrimenti a "non disponibile".
-
- Effetti:
-   Aggiorna il campo 'stato' del veicolo a "disponibile" o "non disponibile"
-   in base alla disponibilità odierna. Restituisce true se il veicolo
-   non è disponibile, false altrimenti.
-*/
-bool aggiorna_stato_veicolo(ptr_veicolo ve) 
-{
-    if (!ve) return false;
-
-    if (veicolo_disponibile_oggi(ve->prenotazione_veicolo)) {
-        strcpy(ve->stato, "disponibile");
-        return false;
-    } else {
-        strcpy(ve->stato, "non disponibile");
-        return true;
-    }
-}
-
-
-/*
- Funzione: libera_veicolo
- ------------------------
- Libera la memoria occupata da un veicolo, comprese le sue prenotazioni.
-
- Parametri:
-   ve: puntatore al veicolo da liberare
+   ve: puntatore al veicolo da modificare.
+   stato: stringa che rappresenta il nuovo stato.
 
  Pre-condizione:
-   ve deve essere un puntatore valido a una struttura veicolo allocata dinamicamente
+   ve deve essere un puntatore valido.
+   stato deve essere una stringa valida.
 
  Post-condizione:
-   La memoria occupata dal veicolo e dalle sue prenotazioni viene liberata
+   Lo stato interno del veicolo è aggiornato con la nuova stringa.
+   La stringa è sempre terminata correttamente.
 
  Effetti:
-   Dealloca la memoria associata al veicolo e alle sue prenotazioni.
+   Modifica il campo stato all’interno della struttura veicolo.
 */
-void libera_veicolo(ptr_veicolo ve)
+void imposta_stato(ptr_veicolo ve, const char *stato)
 {
-    if(ve){
-        libera_prenotazioni(ve->prenotazione_veicolo);
-        free(ve);
+    if (ve && stato) {
+        strncpy(ve->stato, stato, DIM_STATO - 1);
+        ve->stato[DIM_STATO - 1] = '\0'; // garantisce terminazione stringa
     }
 }
